backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyascii==0.3.2
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.0
coqpit==0.0.17
coqui-tts==0.24.2
coqui-tts-trainer==0.1.4
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dateparser==1.1.8
decorator==5.1.1
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
grpcio==1.67.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gunicorn==23.0.0
huggingface-hub==0.26.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.10
networkx==3.2.1
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
packaging==24.1
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
propcache==0.2.0
protobuf==5.28.2
psutil==6.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pygame==2.6.1
Pygments==2.18.0
pyparsing==3.2.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
soundfile==0.12.1
soxr==0.5.0.post1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
SudachiDict-core==20240716
SudachiPy==0.6.8
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.5.0
torchaudio==2.5.0
tqdm==4.66.5
transformers==4.42.4
typeguard==4.3.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.15.5
zipp==3.20.2